Uploaded new hashmaps
Changed HashMapPractice.cpp to work with new hash maps

HashMapD and HashMapS are hashmaps for doubles and strings as keys respectively. Found it difficult to implement the whole this a template so now they're just separate.

Added bool which needs to be initialized by constructor. Determines which hash function to use.

Implemented mid square method as hashA in both and tested. Seems to be working.

Added vector<pair<double/string, int> frequentKeys() for fetching of the most common occuring keys.

When you want to increment a value by one for counting, use the [] operator and ++, like I do in the practice file.